refactor: move worker code
Move worker code into own package, for better encapsulation and testing purposes.
